requirements.txt
amocrm-api==2.6.1
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
PyJWT==2.8.0
requests==2.31.0
urllib3==2.1.0
